chore: pkg build & automate type updates
