--- nanofuzz-package/README.md
@@ -1,7 +1,0 @@
-# NaNofuzz npm package
-
-The npm package nanofuzz-package (we can definitely change the name if needed) contains the type definitions needed to NaNofuzz with custom validators. These type definitions are: FuzzIoElement and FuzzTestResult.
-
-To install the package, first run `npm install nanofuzz-package`.
-
-Then, import the types at the top of the script: `import {FuzzIoElement, FuzzTestResult} from 'nanofuzz-package';`
