--- nanofuzz-package/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "nanofuzz-package",
-  "version": "1.0.2",
-  "description": "Type definitions for nanofuzz",
-  "main": "types.js",
-  "scripts": {
-    "build": "npx tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "nanofuzz-package": "^1.0.2",
-    "typescript": "^4.1.3"
-  }
-}
