+++ packages/runtime/README.md
@@ -1,0 +1,5 @@
+# NaNofuzz runtime support package
+
+This npm package provides the runtime support and types necessary for NaNofuzz tests to be executed outside of Visual Studio Code; e.g., in CI.
+
+To install the package, run `yarn add @nanofuzz/runtime` or `npm install @nanofuzz/runtime`.
