+++ packages/runtime/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@nanofuzz/runtime",
+  "displayName": "NaNofuzz runtime support",
+  "publisher": "penrose",
+  "version": "0.1.15",
+  "description": "Runtime support and types for NaNofuzz",
+  "main": "build/cjs/index.js",
+  "module": "build/esm/index.js",
+  "types": "build/cjs/index.d.ts",
+  "author": "The NaNofuzz Team @ Carnegie Mellon University",
+  "license": "MIT",
+  "scripts": {
+    "checkver": "node checkVer.js",
+    "prebuild": "node prebuild.js",
+    "build": "npx tsc --project tsconfig.cjs.json && npx tsc --project tsconfig.esm.json",
+    "package": "npm publish --dry-run --access public",
+    "publish": "npm publish --access public",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "files": [
+    "build"
+  ],
+  "devDependencies": {
+    "copyfiles": "^2.4.1",
+    "rimraf": "^5.0.5",
+    "typescript": "^4.9.5"
+  }
+}
